+++ src/main/java/io/github/tufkan1/projectex/ProjectEX.java
@@ -2,6 +2,7 @@ package io.github.tufkan1.projectex;
 
 import io.github.tufkan1.projectex.api.emc.EmcValueRegistry;
 import io.github.tufkan1.projectex.command.EmcCommands;
+import io.github.tufkan1.projectex.emc.mapping.minecraft.MinecraftRecipeMappingService;
 import io.github.tufkan1.projectex.emc.reload.EmcDataReloadListener;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.resource.v1.ResourceLoader;
@@ -24,6 +25,7 @@ public final class ProjectEX implements ModInitializer {
             new EmcDataReloadListener(EMC_VALUES)
         );
         EmcCommands.register();
+        MinecraftRecipeMappingService.register();
         LOGGER.info("ProjectEX {} is initializing with {} EMC values", version(), EMC_VALUES.size());
     }